Fix more v8 errors (#336)
- change error message for `Object.create` invalid property descriptor
- disable v8 test cases for deprecated legacy RegExp static properties
  and invalid left hand side error type
- update v8.txt
- fix v8.sh behavior for single tests

+++ v8.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 set -e
 : ${QJS:=build/qjs}
-"$QJS" v8.js $* 2>&1 | tee v8.txt$$
-diff -uw v8.txt v8.txt$$ || exit 1
-rm v8.txt$$
+if [ "x" = "x$1" ] ; then
+  "$QJS" v8.js $* 2>&1 | tee v8.txt$$
+  diff -uw v8.txt v8.txt$$ || exit 1
+  rm v8.txt$$
+else
+  "$QJS" v8.js $* 2>&1
+fi
